svg/styling/css-var-dom-interface.tentative.html


<!DOCTYPE html>
<title>SVGLength DOM interface with CSS variables</title>
<link rel="author" title="Divyansh Mangal" href="mailto:dmangal@microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  :root {
    --length-100px: 100px;
    --length-50: 50;
    --length-75percent: 75%;
    --invalid-length: not-a-length;
  }
</style>
<svg width="200" height="200">
  <rect id="rect-px" width="var(--length-100px)" height="50"/>
  <rect id="rect-number" width="var(--length-50)" height="50"/>
  <rect id="rect-percent" width="var(--length-75percent)" height="50"/>
  <rect id="rect-invalid" width="var(--invalid-length)" height="50"/>
  <rect id="rect-normal" width="100px" height="50"/>
</svg>
<script>
test(() => {
  const rect = document.getElementById('rect-px');
  const width = rect.width.baseVal;
  // SVGLength.value should resolve CSS variable and return computed value
  assert_equals(width.value, 100, 'value should resolve var(--length-100px) to 100');
  // SVGLength.valueInSpecifiedUnits should return 0 for CSS variables
  assert_equals(width.valueInSpecifiedUnits, 0, 'valueInSpecifiedUnits should return 0 for CSS variables');
  // SVGLength.unitType should return SVG_LENGTHTYPE_UNKNOWN for CSS variables
  assert_equals(width.unitType, SVGLength.SVG_LENGTHTYPE_UNKNOWN, 'unitType should be UNKNOWN for CSS variables');
}, 'SVGLength DOM interface with CSS variable containing px unit');
test(() => {
  const rect = document.getElementById('rect-number');
  const width = rect.width.baseVal;
  // CSS variable with number should resolve to user units
  assert_equals(width.value, 50, 'value should resolve var(--length-50) to 50');
  assert_equals(width.valueInSpecifiedUnits, 0, 'valueInSpecifiedUnits should return 0 for CSS variables');
  assert_equals(width.unitType, SVGLength.SVG_LENGTHTYPE_UNKNOWN, 'unitType should be UNKNOWN for CSS variables');
}, 'SVGLength DOM interface with CSS variable containing number');
test(() => {
  const rect = document.getElementById('rect-percent');
  const width = rect.width.baseVal;
  // CSS variable with percentage should resolve to computed value
  assert_equals(width.value, 150, 'value should resolve var(--length-75percent) to 150 (75% of 200px SVG width)');
  assert_equals(width.valueInSpecifiedUnits, 0, 'valueInSpecifiedUnits should return 0 for CSS variables');
  assert_equals(width.unitType, SVGLength.SVG_LENGTHTYPE_UNKNOWN, 'unitType should be UNKNOWN for CSS variables');
}, 'SVGLength DOM interface with CSS variable containing percentage');
test(() => {
  const rect = document.getElementById('rect-invalid');
  const width = rect.width.baseVal;
  assert_equals(width.value, 0, 'parsing failed for invalid CSS variable');
  assert_equals(width.valueInSpecifiedUnits, 0, 'valueInSpecifiedUnits should return 0 for CSS variables');
  assert_equals(width.unitType, SVGLength.SVG_LENGTHTYPE_UNKNOWN, 'unitType should be UNKNOWN for CSS variables');
}, 'SVGLength DOM interface with invalid CSS variable');
test(() => {
  const rect = document.getElementById('rect-normal');
  const width = rect.width.baseVal;
  // Normal length without CSS variable for comparison
  assert_equals(width.value, 100, 'value should return 100 for normal 100px length');
  assert_equals(width.valueInSpecifiedUnits, 100, 'valueInSpecifiedUnits should return 100 for normal length');
  assert_equals(width.unitType, SVGLength.SVG_LENGTHTYPE_PX, 'unitType should be PX for normal px length');
}, 'SVGLength DOM interface with normal length (no CSS variables) for comparison');
test(() => {
  const rect = document.getElementById('rect-px');
  const width = rect.width.baseVal;
  // Test that DOM interface methods work with CSS variables
  assert_equals(width.valueAsString, 'var(--length-100px)', 'valueAsString should preserve CSS variable syntax');
  // Setting value should work (converts to user units)
  width.value = 120;
  assert_equals(width.value, 120, 'setting value should work');
  assert_not_equals(width.valueAsString, 'var(--length-100px)', 'valueAsString should no longer be CSS variable after setting value');
  // Reset for further tests
  rect.setAttribute('width', 'var(--length-100px)');
  assert_equals(rect.width.baseVal.valueAsString, 'var(--length-100px)', 'reset should work');
}, 'SVGLength DOM interface methods with CSS variables');
test(() => {
  const rect = document.getElementById('rect-px');
  const width = rect.width.baseVal;
  // Test convertToSpecifiedUnits behavior with CSS variables
  assert_equals(width.valueAsString, 'var(--length-100px)', 'initial value should be CSS variable');
  width.convertToSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_CM);
  // After conversion, the value should no longer be a CSS variable
  assert_not_equals(width.valueAsString, 'var(--length-100px)', 'value should no longer be CSS variable after conversion');
  assert_equals(width.unitType, SVGLength.SVG_LENGTHTYPE_CM, 'unitType should be CM after conversion');
  assert_approx_equals(width.value, 100, 1, 'value should be approximately the same after unit conversion');
  // Reset for further tests
  rect.setAttribute('width', 'var(--length-100px)');
  assert_equals(rect.width.baseVal.valueAsString, 'var(--length-100px)', 'reset should work');
}, 'SVGLength.convertToSpecifiedUnits with CSS variables');
test(() => {
  const rect = document.getElementById('rect-px');
  const width = rect.width.baseVal;
  // Test newValueSpecifiedUnits behavior with CSS variables
  width.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_CM, 2.54);
  assert_equals(width.value, 96, 'newValueSpecifiedUnits should replace CSS variable with specified value');
  assert_equals(width.unitType, SVGLength.SVG_LENGTHTYPE_CM, 'unitType should change after newValueSpecifiedUnits');
  assert_not_equals(width.valueAsString, 'var(--length-100px)', 'valueAsString should no longer be CSS variable');
}, 'SVGLength.newValueSpecifiedUnits with CSS variables');
</script>
